Extract shared module style hook
